requirements.txt
fastapi==0.88.0
neo4j==5.3.0
pandas==1.5.0
py2neo==2021.2.3
pydantic==1.10.2
pymongo==4.3.2
python-dotenv==0.21.0
python_dateutil==2.8.2
pytz==2022.2.1
uvicorn==0.20.0
openpyxl==3.0.10